+++ packages/otfjs-ui/src/components/font-list/font-list.components/font-grid.tsx
@@ -1,13 +1,13 @@
 import { memo, useEffect, useMemo, useRef } from 'react'
 import { Link, useRouter } from '@tanstack/react-router'
 
-import { useTimeoutAfterSet } from '../../hooks/use-timeout-after-set'
-import { handle } from '../../shortcuts/shortcuts'
-import { Fonts } from '../../types/fonts'
-import { createElementWalkerFactory } from '../../utils/dom'
-import { addListener } from '../../utils/event'
-import { entriesFilterMap } from '../../utils/object'
-import { FontIcon } from '../font-icon/font-icon'
+import { useTimeoutAfterSet } from '../../../hooks/use-timeout-after-set'
+import { handle } from '../../../shortcuts/shortcuts'
+import { Fonts } from '../../../types/fonts'
+import { createElementWalkerFactory } from '../../../utils/dom'
+import { addListener } from '../../../utils/event'
+import { entriesFilterMap } from '../../../utils/object'
+import { FontIcon } from '../../font-icon/font-icon'
 
 import styles from './font-grid.module.css'
