+++ packages/otfjs-ui/src/components/font-list/font-list.tsx
@@ -1,10 +1,10 @@
 import { useDeferredValue, useState } from 'react'
 
 import fonts from '../../fonts.json'
-import { SearchBar } from './components/search-bar'
-import { FontGrid } from './font-grid'
+import { FontGrid } from './font-list.components/font-grid'
+import { SearchBar } from './font-list.components/search-bar'
 
-import styles from './no-font-view.module.css'
+import styles from './font-list.module.css'
 
 export function NoFontView() {
   const [filter, setFilter] = useState('')
