+++ packages/otfjs-ui/src/components/font-view/components/head.module.css
@@ -1,8 +1,9 @@
 .head {
+  grid-area: head;
   display: grid;
   grid-template-columns: auto 1fr;
   height: 97px;
   gap: 8px;
   padding: 8px;
-  border-bottom: 1px solid var(--color-border);
+  background: var(--color-bg-sidebar);
 }
