+++ packages/otfjs-ui/src/components/font-view/font-view.module.css
@@ -1,19 +1,30 @@
 .root {
   display: grid;
-  grid-template-columns: auto 1fr;
+  gap: 1px;
+  grid:
+    'head view' auto
+    'tabs view' 1fr / 240px 1fr;
   height: 100%;
   overflow: hidden;
-}
 
-.sidebar {
-  overflow: scroll;
-  width: 240px;
-  box-sizing: content-box;
-  border-right: 1px solid var(--color-border);
-  background: var(--color-bg-sidebar);
+  background: var(--color-border);
+
+  :where(& > *) {
+    background: var(--color-bg);
+  }
+
+  @media screen and (max-width: 600px) {
+    grid:
+      'head head' auto
+      'tabs view' 1fr / 100px 1fr;
+  }
 }
 
 .tabs {
+  grid-area: tabs;
+  overflow: scroll;
+  background: var(--color-bg-sidebar);
+
   & ul {
     list-style: none;
     padding: 0;
@@ -50,6 +61,7 @@
 }
 
 .tableView {
+  grid-area: view;
   position: relative;
   overflow: scroll;
 }
