+++ packages/otfjs-ui/src/routes/index.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, useRouterState } from '@tanstack/react-router'
 
+import { NoFontView } from '../components/font-list/font-list'
 import { FontView } from '../components/font-view/font-view'
-import { NoFontView } from '../components/no-font-view/no-font-view'
 import { getFontById } from '../utils/fetch-font'
 
 export const Route = createFileRoute('/')({
